+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/locationresource/LocationResourcePrisonIdTest.kt
@@ -4,9 +4,8 @@ import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import uk.gov.justice.digital.hmpps.locationsinsideprison.integration.CommonDataTestBase
+import uk.gov.justice.digital.hmpps.locationsinsideprison.resource.EXPECTED_USERNAME
 import uk.gov.justice.hmpps.test.kotlin.auth.WithMockAuthUser
-
-const val EXPECTED_USERNAME = "A_TEST_USER"
 
 @WithMockAuthUser(username = EXPECTED_USERNAME)
 class LocationResourcePrisonIdTest : CommonDataTestBase() {
